Adicionando get para enderecoController
Adicionando get para enderecoController
